fix: move result class to a package where proguard is not involved to make gson happy
