+++ app/src/main/java/ethiopia/covid/android/ui/fragment/ResultFragment.java
@@ -29,6 +29,7 @@ import java.util.Map;
 import ethiopia.covid.android.R;
 import ethiopia.covid.android.data.QuestionItem;
 import ethiopia.covid.android.data.QuestionnaireItem;
+import ethiopia.covid.android.data.Result;
 import ethiopia.covid.android.ui.activity.MainActivity;
 import ethiopia.covid.android.ui.adapter.ResultRecyclerAdapter;
 import timber.log.Timber;
@@ -48,52 +49,6 @@ public class ResultFragment extends BaseFragment {
     private MaterialButton sendButton;
 
     private FirebaseFirestore firestore = FirebaseFirestore.getInstance();
-
-    private static class Result {
-        private Map<QuestionnaireItem, List<QuestionItem>> questionItems;
-        private double longitude;
-        private double latitude;
-        private double accuracy;
-
-        public Result(Map<QuestionnaireItem, List<QuestionItem>> questionItems, double longitude, double latitude, double accuracy) {
-            this.questionItems = questionItems;
-            this.longitude = longitude;
-            this.latitude = latitude;
-            this.accuracy = accuracy;
-        }
-
-        public Map<QuestionnaireItem, List<QuestionItem>> getQuestionItems() {
-            return questionItems;
-        }
-
-        public void setQuestionItems(Map<QuestionnaireItem, List<QuestionItem>> questionItems) {
-            this.questionItems = questionItems;
-        }
-
-        public double getLongitude() {
-            return longitude;
-        }
-
-        public void setLongitude(double longitude) {
-            this.longitude = longitude;
-        }
-
-        public double getLatitude() {
-            return latitude;
-        }
-
-        public void setLatitude(double latitude) {
-            this.latitude = latitude;
-        }
-
-        public double getAccuracy() {
-            return accuracy;
-        }
-
-        public void setAccuracy(double accuracy) {
-            this.accuracy = accuracy;
-        }
-    }
 
     private ResultFragment(Map<QuestionnaireItem, List<QuestionItem>> questionItems, String hashOfQuestionnaire) {
         this.questionItems = questionItems;
